requirements-data.txt
kaggle>=1.6.0
requests>=2.31.0
Pillow>=10.3.0
pandas>=2.2.0
opencv-python-headless>=4.9.0
datasets>=2.19.0
huggingface_hub>=0.23.0
PyYAML>=6.0.1
tqdm>=4.66.0
scikit-learn>=1.4.0
openpyxl>=3.1.2
